Fixed bug where streams would be inappropriately closed.
`Stream.open` used to return the underlying stream as is, and so it
would get closed when used in a context manager.  But this doesn't
reflect what would likely be the intended scope.  That is, a stream open
before a context manager should still be open afterwards.  In other
words, the context manager encloses a local scope which is expected to
be isolated from the outer scope.

The immediate solution is to wrap the stream in a context manager that
doesn't close the stream, but the long-term solution is to wrap the
stream in a file-like object that doesn't close its underlying stream
when it's closed.

* Tested with a few types of open file-like objects.