Separated parsing of constants from parsing their common synonyms.
Because `True`, `False`, and `None` often need to get recognized
specifically, distinctly from their synonyms.  This gives control to the
user whether to "enable" synonyms or not.  Also, sets of strings (like
enumerations or categorical variables) usually need to be recognized /
parsed in the context of the entire set of values, so individually
pattern-matching words can easily go astray.  For example, should `'on'`
and `'no'` be converted to `True` and `False` or left as is?  (Before
this change, these words would have been treated as synonyms without
regard for context like {on, in, under, over} or {no, oc, fc, uc}.)