Deleted experiment with convolutions.  Look at previous to resurrect.
Trying to do density of real values via convolutions ended up being too
much of a rabbit hole with questionable benefit, so I'm abandoning that
(and giving up on generalization and reusability) in favor of a
concrete, specific implementation.  If convolutions later make sense for
the implementation, then I'll come back to this code at that point with
the specific implementation as a guide for what to do.