Reimplemented 2-by-2 tables based on 'contingency_tables.py'.
* Implemented new scores.

A version of this commit was originally made at
2024-10-09T17:43:26-05:00 on a branch that combined
'contingency_tables.py' and 'classification_analysis.py', but that was a
nightmare for backwards compatibility, so I abandoned that in favor of
keeping them separate.  I would have preferred to cherry pick the
commit, but that didn't work because the files were mixed together, so
we have this new commit.